==> 1_ComputeRelativeHeights.ipynb <==
# Compute relative heights 


This file reads in the entire cleaned dataset with all labels and then creates an additional column with the relative heights 

import numpy as np
import pandas as pd

# Define path variables and file name  
FILE_PATH = r"../DATA"
FILE_NAME_INPUT = r"/insert_file_name_here.txt"
FILE_NAME_OUTPUT = r"/insert_file_name_here.csv"
df = pd.read_csv(FILE_PATH+FILE_NAME_INPUT, delimiter=',')

# Rename the column 
df = df.rename(index=str, columns={"//X": "X"})
# Obtain number of observations
n = df.iloc[0,0]
#Print number of instances as control value
print "number of points in dataset", int(n)
# Get rid of first line with number of observations
df = df.drop(df.index[0])

# Test whether there are any NAN values in the data set 
if df.isnull().values.any() == False:
    print("No Nan in this data set")
else:
    print("There are Nan in this dataset")

# Create a new column with relative height 
# Positive rel_height if the point is higher, and negative rel_height if point is lower
# than the scanner
df['relative_height']=df['Z'] - df['z_origin']
# Save file
df.to_csv(FILE_PATH+FILE_NAME_OUTPUT,delimiter=',')
