==> 3_CreateClassesForClassification.ipynb <==
# Create classes to be used during the ML process 

**IMPORTANT**
The file generated in 2_AnalyseClasses (last line of code) was assessed in CloudCompare and cleaned manually. Therefore, it is crucial to use the cleaned dataset generated in CloudCompare here. 

Cleaned: varying PC density, inhomogeneous classes, etc 

import numpy as np
import pandas as pd
import os
import matplotlib.pyplot as plt

# Define a data frame with all my data# Define  
FILE_PATH = r"../DATA"
# For input data
FILE_NAME = r"/insert_name_here.txt"
# Input data from CloudCompare with computed surface normals
FILE_NAME2 = r"/insert_name_here.txt"
# For re-labelled class file to be exported and then compute surface normals on that
FILE_NAME_SN = r"/insert_name_here.txt"
# Output file name for file containing surface normals and verticality (last line of code)
FILE_NAME_V = r"/insert_name_here.txt"
IMAGE_FILE_PATH = r"images"
df = pd.read_csv(FILE_PATH+FILE_NAME, delimiter=',')

# Obtain number of observations
n = df.iloc[0,0]
n = int(n)
# Get rid of first line with number of observations
df = df.drop(df.index[0])
print n

# Rename the column 
df = df.rename(index=str, columns={"//X": "X"})
df = df.rename(index=str, columns={"Scalar field": "x_origin"})
df = df.rename(index=str, columns={"Scalar field #2": "y_origin"})
df = df.rename(index=str, columns={"Scalar field #3": "z_origin"})
df = df.rename(index=str, columns={"Scalar field #4": "reflectance"})
df = df.rename(index=str, columns={"Scalar field #5": "theta"})
df = df.rename(index=str, columns={"Scalar field #6": "id"})
df = df.rename(index=str, columns={"Scalar field #7": "class"})
df = df.rename(index=str, columns={"Scalar field #8": "range"})
df = df.rename(index=str, columns={"Scalar field #9": "relative_height"})
df = df.rename(index=str, columns={"Scalar field #10": "revised_class"})
df = df.rename(index=str, columns={"Scalar field #11": "revised_class_1_15"})

# Convert the two columns into integer values
df = df.astype({"id": 'int64', "class": 'int64', "revised_class": 'int64', "revised_class_1_15": 'int64'})

df.head()

## After visual inspection
The following steps were undertaken after visual inspection in CloudCompare 

Combine classes and delete empty classes

# Delete class 10 
df["revised_class_1_15"].replace(10,np.NaN, inplace = True)
df.head()

# Test whether there are any NAN values in the data set 
if df.isnull().values.any() == False:
    print("No Nan in this data set")
else:
    print("There are Nan in this dataset")
    df.dropna(inplace = True)
    print("NaNs were deleted.")
    
# Convert the class labels back into integer values (because of NaN it became float)
df = df.astype({"revised_class_1_15": 'int64'})

# Merge classes 1 and 2 (road and other ground)
df["revised_class_1_15"].where(df["revised_class_1_15"] != 1, 
                                   2, inplace=True)

## Create classes for ML 
This involves:
    
Delete all class columns other than the most recent one

Delete ID (only relevant for segmentation) 

Keeping only the classes 1-11 (revised_class_1_15) - 15 before, but some deleted if empty after cleaning in CloudCompare

Rename the most recent class column to *class*

columns_to_drop = ["x_origin", "y_origin", "z_origin", "id", "class", "revised_class"]

for elem in columns_to_drop:
    df.drop([elem], axis=1, inplace=True)
    
df.rename(index=str, columns={"revised_class_1_15": "class"}, inplace = True)
df.head()

# Save file to be exported for CloudCompare. 
# This file should be used to compute surface normals on. 
df.to_csv(FILE_PATH+, index=False)

# Prepare ML dataset with surface normals 

I exported the dataset file just created above and then I computed the surface normals in Cloud Compare. The file with the surface normals is called Dataset_for_ML_SurfaceNormals.txt (Choose any name)

Now, this file loaded in here, contains the revised classes, and the surface normals (nx,ny and nz). This needs to be analysed now. 

df2 = pd.read_csv(FILE_PATH+FILE_NAME2, delimiter=',')

# Rename the column 
df2 = df2.rename(index=str, columns={"//X": "X"})
df2 = df2.rename(index=str, columns={"Scalar field": "reflectance"})
df2 = df2.rename(index=str, columns={"Scalar field #2": "theta"})
df2 = df2.rename(index=str, columns={"Scalar field #3": "range"})
df2 = df2.rename(index=str, columns={"Scalar field #4": "relative_height"})
df2 = df2.rename(index=str, columns={"Scalar field #5": "class"})

df2.head()

# Now compute the verticality (Formula given in report)
df2['verticality'] = 1 - df2['Nz']

# Only keep the relevant columns
columns_to_drop = ["Nx", "Ny", "Nz", "theta", "range"]

for elem in columns_to_drop:
    df2.drop([elem], axis=1, inplace=True)
    
# Convert the two columns ID and CLASS into integer values
df2 = df2.astype({"class": 'int64'})

# Save file
df2.to_csv(FILE_PATH+FILE_NAME_V, index=False)

# Clear memory 
df = None
df2 = None
