==> 6_ScenarioB_calcfeatures.ipynb <==
# Machine Learning now

import pandas as pd
import os
import numpy as np
import sklearn 
import scipy.linalg as scplinag
from sklearn.neighbors import KDTree
from scipy.spatial import distance

# Scenario B

Take the same features as in Scenario 1, add all of them in the same high-dim vector and add the reflectance as well. So we can see if that changes anything. 

Here, we just create the feature vector, wait for new optimised results for RF. 

# Define a data frame with all my data# Define  
FILE_PATH = r"../DATA"
FILE_PATH_ScenarioB = r"../DATA/ML_datasets/ScenarioB"

FILE_NAME_50_ref = r"/ML_dataset_50NN_reflectance_withXYZ.txt"
FILE_NAME_10 = r"/ScenarioB_10NNFeatures.txt"
FILE_NAME_100 = r"/ScenarioB_100NNFeatures.txt"

# Read in file with 50NN and reflectance this will be the base file
# call it df1
df_main = pd.read_csv(FILE_PATH_ScenarioB+FILE_NAME_50_ref, delimiter=',')

# Now read in the files for 10 and 100 NN features  
df_10 = pd.read_csv(FILE_PATH_ScenarioB+FILE_NAME_10, delimiter=',')
df_100 = pd.read_csv(FILE_PATH_ScenarioB+FILE_NAME_100, delimiter=',')

# Need to rename the attributes according to their NN size
Here, we do it manually because each set is very different, all we do is indicate the attributes whether they belong to 10NN or other NN

### Rename for 50 NN 

# Rename the column 
df_main = df_main.rename(index=str, columns={"lambda1": "lambda1_50NN"})
df_main = df_main.rename(index=str, columns={"lambda2": "lambda2_50NN"})
df_main = df_main.rename(index=str, columns={"lambda3": "lambda3_50NN"})
df_main = df_main.rename(index=str, columns={"lambda4": "lambda4_50NN"})
df_main = df_main.rename(index=str, columns={"lambda5": "lambda5_50NN"})
df_main = df_main.rename(index=str, columns={"lambda6": "lambda6_50NN"})
df_main = df_main.rename(index=str, columns={"lambda7": "lambda7_50NN"})
df_main = df_main.rename(index=str, columns={"lambda8": "lambda8_50NN"})
df_main = df_main.rename(index=str, columns={"local_density": "local_density_50NN"})
df_main = df_main.rename(index=str, columns={"verticality": "verticality_50NN"})

df_main.head()

list_10 = list(df_10.columns.values)
list_100 = list(df_100.columns.values)
print list_100

df_10.head()

df_main['lambda1_10NN'] = df_10.iloc[:,0].values

df_main.head()

for count, elem in enumerate(list_10):
    df_main[elem] = df_10.iloc[:,count].values

df_main.head()

for count, elem in enumerate(list_100):
    df_main[elem] = df_100.iloc[:,count].values

I just created my scenario B file with all attributes, this is the size: (11052294, 33)

# Now I have to create my training and testing dataset

df_10 = None
df_100 = None

# Testing and training 

from sklearn.model_selection import StratifiedShuffleSplit

df_main.head()

data = df_main.drop(columns=["class"])
X = data.values
y = df_main["class"].values

print data.shape
print X.shape
print y.shape

split = StratifiedShuffleSplit(n_splits=10, test_size=0.2, random_state=42)
for train_index, test_index in split.split(X, y):
    # Do that here
    X_train, X_test = X[train_index], X[test_index]
    y_train, y_test = y[train_index], y[test_index]

print "Size X training data", X_train.shape
print "Size y training data", y_train.shape
print "Size X testing data", X_test.shape
print "Size y testing data", y_test.shape

# Now save everything

# Save all my data as NUMPY arrays
# To load them later
# np.loadtxt()
#np.savetxt(FILE_PATH_ScenarioB+'/y_train.txt', y_train, delimiter=',')
np.savetxt(FILE_PATH_ScenarioB+'/y_test_withXYZ.txt', y_test, delimiter=',')
#np.savetxt(FILE_PATH_ScenarioB+'/X_train.txt', X_train, delimiter=',')
np.savetxt(FILE_PATH_ScenarioB+'/X_test_withXYZ.txt', X_test, delimiter=',')

# Clean variables I domnt need right now
df = None
corr_matrix = None
X = None
y = None
data = None 
df_main = None
df_10 = None
