==> 6_scenarioB_createfeatures10NN.ipynb <==
# Machine Learning now

import pandas as pd
import os
import numpy as np
import sklearn 
import scipy.linalg as scplinag
from sklearn.neighbors import KDTree
from scipy.spatial import distance

# Whole dataset and 50 NN 

Attributes as in Weinmann, 2013, plus relative height (Weinmann, 2014) 

# Define a data frame with all my data# Define  
FILE_PATH_ScenarioB = r"../DATA/ML_datasets/ScenarioB"
FILE_PATH = r"../DATA"
FILE_NAME = r"/5_Data_ML_attributes_10NN.txt"
IMAGE_FILE_PATH = r"images"
df = pd.read_csv(FILE_PATH+FILE_NAME, delimiter=',')
df.head()

columns_to_drop_10 = ["X", "Y", "Z", "radius_neighbourhood", "relative_height"]

for elem in columns_to_drop_10:
    df.drop([elem], axis=1, inplace=True)

df.head()

### I have made a mistake with the class names, therefore I need to re-name them now

In the file 3_CreateFinalClasses, I did not change the names of the classes properly. Therefore, I will have to re-name them now. 

Manually, not to confuse it.

# Road (second most points)
df["class"].where(df["class"] != 2, 1, inplace=True)
# Sidewalk
df["class"].where(df["class"] != 3, 2, inplace=True)
# Curb
df["class"].where(df["class"] != 4, 3, inplace=True)
# Building (most points)
df["class"].where(df["class"] != 5, 4, inplace=True)
# Other pole like objects
df["class"].where(df["class"] != 6, 5, inplace=True)
# Small poles
df["class"].where(df["class"] != 7, 6, inplace=True)
# Pedestrians
df["class"].where(df["class"] != 11, 7, inplace=True)
# 2 wheelers
df["class"].where(df["class"] != 12, 8, inplace=True)
# 4 wheelers
df["class"].where(df["class"] != 13, 9, inplace=True)
# Trees
df["class"].where(df["class"] != 14, 10, inplace=True)
# Potted plants
df["class"].where(df["class"] != 15, 11, inplace=True)

df['class'].value_counts()

df.drop(["class"], axis=1, inplace=True)
df.head()

# Rename the columns to make it easier

df = df.rename(index=str, columns={"lambda1": "lambda1_10NN"})
df = df.rename(index=str, columns={"lambda2": "lambda2_10NN"})
df = df.rename(index=str, columns={"lambda3": "lambda3_10NN"})
df = df.rename(index=str, columns={"lambda4": "lambda4_10NN"})
df = df.rename(index=str, columns={"lambda5": "lambda5_10NN"})
df = df.rename(index=str, columns={"lambda6": "lambda6_10NN"})
df = df.rename(index=str, columns={"lambda7": "lambda7_10NN"})
df = df.rename(index=str, columns={"lambda8": "lambda8_10NN"})
df = df.rename(index=str, columns={"local_density": "local_density_10NN"})
df = df.rename(index=str, columns={"verticality": "verticality_10NN"})

df.head()

## Now save that in scenario B folder

df.to_csv(FILE_PATH_ScenarioB+'/ScenarioB_10NNFeatures.txt', index= False)

# Clean variables I domnt need right now

df = None
corr_matrix = None
X = None
y = None
data = None
