==> 7_ML.ipynb <==
# Compute additional features

My initial features for ML are:

8-feature descriptor, local point density, verticality, relative height  

Now add: 
Maybe follow Schindler, stack them together, maybe reflectivity, theta 

import numpy as np
import sklearn 
from sklearn.neighbors import KDTree
from sklearn.ensemble import RandomForestClassifier
import time
from sklearn import metrics

# Define a data frame with all my data# Define  
FILE_PATH = r"../DATA/ML_datasets/Initial_setup"
META_FILE_PATH = "../DATA/META"
IMAGE_FILE_PATH = r"images"

y_train = np.loadtxt(FILE_PATH+'/y_train_100NN.txt', delimiter=',')
X_train = np.loadtxt(FILE_PATH+'/X_train_100NN.txt', delimiter=',')

# For RF classifier 
y_train = y_train.astype('float32')
X_train = X_train.astype('float32')

# Work with small subset for now, change that later
y_train_subset = y_train[:10000]
X_train_subset = X_train[:10000]

start = time.time()
clf = RandomForestClassifier(max_depth=4, random_state=42, n_estimators=100, criterion='gini')
print 'Created Random Forest in:', float(time.time()-start), 'seconds'
start = time.time()
clf.fit(X_train_subset, y_train_subset)
print 'Fit model in:', float(time.time()-start), 'seconds'

## Now make predictions

Predict and look at probabilities

test_predictions = X_train_subset[:4]
test_predictions_labels = y_train_subset[:4]

for count, elem in enumerate(test_predictions):
    print "Predicted value:", clf.predict([elem])
    print "True label:", test_predictions_labels[count]
    print "Predicted with ", clf.predict_proba([elem]), "probability"

# This is a Numpy array with all predictions for the whole dataset
y_predictions = clf.predict(X_train_subset)

## Now evaluate the model 

From here on, evaluate the classifier we just created 

# Mean accuracy of the dataset 
# average accuracy is the average of each accuracy per class 
# (sum of accuracy for each class predicted/number of class)
print "Mean accuracy:", clf.score(X_train_subset, y_train_subset)

# Determine accuracy score
# http://scikit-learn.org/stable/modules/model_evaluation.html#accuracy-score

from sklearn.metrics import accuracy_score
print "Accuracy score", accuracy_score(y_train_subset, y_predictions)

# This gives me the ttal number of instances that were classified correctly 
# If I divide that by the number of all instances, I get the first result
# return the number of correctly classified samples
print "Accuracy score (not normalized)", accuracy_score(y_train_subset, y_predictions,normalize=False)

# Of course that accuracy will be high because there are a lot of instances for building and road, so when 
# we are looking at all instances (not single classes), we expect a good result

# Determine average precision score
# NOT: ONLYAVAILABLE FOR BINARY OR MULTILABEL 
# from sklearn.metrics import average_precision_score
# average_precision_score(y_train_subset, y_predictions)

# Classification report
# http://scikit-learn.org/stable/modules/generated/sklearn.metrics.classification_report.html#sklearn.metrics.classification_report

from sklearn.metrics import classification_report
clf_report = classification_report(y_train_subset, y_predictions)
clf_report = clf_report.encode('ascii','ignore')
print clf_report

# Save it in new text file
file = open(META_FILE_PATH + "/clf_report_100NN.txt", "w")
file.write(clf_report)
file.close()

# Compute F1 score
# http://scikit-learn.org/stable/modules/generated/sklearn.metrics.f1_score.html#sklearn.metrics.f1_score
from sklearn.metrics import f1_score
print "Weighted F1 score:", "%.4f" % f1_score(y_train_subset, y_predictions, average='weighted')
print "Macro F1 score:", "%.4f" % f1_score(y_train_subset, y_predictions, average='macro')
print "Micro F1 score:", f1_score(y_train_subset, y_predictions, average='micro')

# Compute recall
# http://scikit-learn.org/stable/modules/generated/sklearn.metrics.recall_score.html#sklearn.metrics.recall_score

from sklearn.metrics import recall_score
print "Macro RECALL score:", ", %.4f" % recall_score(y_train_subset, y_predictions, average='macro')
print "Micro RECALL score:", ", %.4f" % recall_score(y_train_subset, y_predictions, average='micro')
print "Weighted RECALL score:", "%.4f" % recall_score(y_train_subset, y_predictions, average='weighted')
print "RECALL for each class with average = None", recall_score(y_train_subset, y_predictions, average=None)

# Compute precision
# http://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_score.html#sklearn.metrics.precision_score

from sklearn.metrics import precision_score
print "Precision score (Macro):", "%.4f" % precision_score(y_train_subset, y_predictions, average='macro') 
print "Precision score (Micro):", "%.4f" % precision_score(y_train_subset, y_predictions, average='micro')
print "Precision score (Weighted):", "%.4f" % precision_score(y_train_subset, y_predictions, average='weighted')


# Precision, recall, F1 and support 
# http://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_recall_fscore_support.html#sklearn.metrics.precision_recall_fscore_support
from sklearn.metrics import precision_recall_fscore_support

print "Precision-Recall-Fscore support (Macro):", precision_recall_fscore_support(y_train_subset, y_predictions, average='macro')
print "Precision-Recall-Fscore support (Micro):", precision_recall_fscore_support(y_train_subset, y_predictions, average='micro')
print "Precision-Recall-Fscore support (Weighted):", precision_recall_fscore_support(y_train_subset, y_predictions, average='weighted')

# Get classifier parameters 
clf.get_params

# Feature importance:
# How important each feature is for making a prediction 

feat_importance = clf.feature_importances_
feat_names = ['Linearity', 'Planarity', 'Scattering', 'Omnivariance', 'Anisotropy',
              'Eigentropy', 'Sum of eigenvalues','Change of curvature', 'Local density', 'Relative height', 'Verticality']

print "FEATURE IMPORTANCE:"
for count, elem in enumerate(feat_importance):
    print feat_names[count], "%.4f" % elem

## Crossvalidation 

We have the training and the test dataset for each NN(10, 50, 100). We need another validation set. This could be a small part of the training set: or better use cross validation. (Book 69)

**k-fold cress-validation:** randomly splits the training dataset into (e.g. 10) distinct subsets (folds), then it trains and evaluates the model 10 times. It picks 9 folds for training and one fold for evaluating (validation set), and picks different combinations each time). The result is an array with 10 evaluation scores. 

from sklearn.model_selection import cross_val_score

scores = cross_val_score(clf, X_train_subset, y_train_subset, cv = 3)
print scores

# Note that the score by default is the score parameter specified for the estimator 
# For random forest: 
# Returns the mean accuracy on the given test data and labels.

# http://scikit-learn.org/stable/modules/cross_validation.html#cross-validation

# The mean score and the 95% confidence interval of the score estimate are hence given by: 
print("Accuracy: %0.5f (+/- %0.2f)" % (scores.mean(), scores.std() * 2))

# However, I can just pick a score I want and then instead of calculating it once like above, 
# I would always do cross-validation 
scores = cross_val_score(clf, X_train_subset, y_train_subset, cv = 3, scoring='f1_macro')
print scores

# All scores available here:
# http://scikit-learn.org/stable/modules/model_evaluation.html#scoring-parameter 

X = None
y = None
X_train = None
y_train = None

## Save a classifier and then call it later 

from sklearn.externals import joblib
joblib.dump(clf, 'classifier_100NN.pkl')
# Load job:
# clf = joblib.load('filename.pkl')
