==> 7_Scenario_A2.ipynb <==
# Down-sampling

For scenario A.2 this was required

import numpy as np
import sklearn 
from sklearn.neighbors import KDTree
from sklearn.ensemble import RandomForestClassifier
import time
from sklearn import metrics
from sklearn.externals import joblib

# Define a data frame with all my data# Define  
FILE_PATH = r"../DATA/ML_datasets/Initial_setup"
META_FILE_PATH = "../DATA/META"
IMAGE_FILE_PATH = r"images"

# Important: insert here the number of nearest neighbours
y_train = np.loadtxt(FILE_PATH+'/y_train_xxNN.txt', delimiter=',')
X_train = np.loadtxt(FILE_PATH+'/X_train_xxNN.txt', delimiter=',')

# For RF classifier: training data
y_train = y_train.astype('float32')
X_train = X_train.astype('float32')

### Downsampling
Here, the smallest class is determined and all other classes are downsampled accordingly 

# Create an empty matrix and fill it with the features and then label at the end 
M = np.zeros((X_train.shape[0], X_train.shape[1]+1))
M[:,-1]=y_train
M[:,:11] = X_train

LOT_classes = range(1,12)
smallest_class_value = 3000

data = []
for class_label in LOT_classes:
    N = M[np.where(M[:,-1]==class_label)]
    S = N[:smallest_class_value,:]
    data.append(S)

data_array = np.array(data)
print "Dimensions of array of all data", data_array.shape

data = data_array.reshape((data_array.shape[0] *smallest_class_value, data_array.shape[2]))
print "Real shape for the training data", data.shape

print "First instance before shuffling:", data[0]
np.random.shuffle(data)
print "First instance after shuffling:", data[0]

# Save all my data as NUMPY arrays
# To load them later
# np.loadtxt()
np.savetxt(FILE_PATH+'/y_train_xxNN_3000samples.txt', data[:,-1], delimiter=',')
np.savetxt(FILE_PATH+'/X_train_xxNN_3000samples.txt', data[:,:11], delimiter=',')

X_train_downsample = data[:,:11]
y_train_downsample = data[:,-1]

### Classification for scenario A.2
This follows Weinmann (2014)

start = time.time()
clf = RandomForestClassifier(max_depth=4, random_state=42, n_estimators=100, criterion='gini')
print 'Created Random Forest in:', float(time.time()-start), 'seconds'
start = time.time()
clf.fit(X_train_downsample, y_train_downsample)
print 'Fit model in:', float(time.time()-start), 'seconds'

# saving a classifier
filename = FILE_PATH+'/randomforest_model_xxNN_downsample.sav'
# instead of classifier, rename based on your file
joblib.dump(clf, filename)
