==> 9_EvaluateClassifier2.ipynb <==
import numpy as np
import sklearn 
from sklearn.neighbors import KDTree
from sklearn.ensemble import RandomForestClassifier
import time
from sklearn import metrics
import pandas as pd
from sklearn.externals import joblib

# Define NN size 
NN = 100

## Load in all data

After calculating the overall accuracy, I will not work with the training dataset anymore, but only test it on the testing dataset. 

# Uploadng my datasets
# Define a data frame with all my data# Define  
FILE_PATH = r"../DATA/ML_datasets/Initial_setup"
META_FILE_PATH = "../DATA/META"
IMAGE_FILE_PATH = r"images"

# Testing data
y_test = np.loadtxt(FILE_PATH+'/y_test_'+ str(NN) +'NN.txt', delimiter=',')
X_test = np.loadtxt(FILE_PATH+'/X_test_'+ str(NN) +'NN.txt', delimiter=',')

# For RF classifier: cast as float32 
y_test = y_test.astype('float32')
X_test = X_test.astype('float32')

###uploading the saved classifier
filename = '/randomforest_model_'+ str(NN) +'NN.sav'
clf = joblib.load(FILE_PATH + filename)

# Load the predictions
predictions_train = np.loadtxt(FILE_PATH+'/predictions_train_'+ str(NN) +'NN.txt', delimiter=',')
predictions_test = np.loadtxt(FILE_PATH+'/predictions_test_'+ str(NN) +'NN.txt', delimiter=',')

## Now start evaluation 

### Show problem with threshold or unclassified

I thought I could determine a threshold or a difference between two values to create a class unclassified. But this will not work. Especially the predictions for the noisy and small classes are really bad. They are never the highest in the probability array. Mostly, they get classified as building. Even if I set a threshold, the classsifier is still sure that it must be a building. Also, there is a biggap to the next highest value. 

# Calculate all probabilities for the testing set 
probability_array = clf.predict_proba(X_test)
print "This is the shape of the array", probability_array.shape

probability_array.sort()
print "Max", (probability_array[0][-1])*100, "%"
print "sec Max", (probability_array[0][-2])*100, "%"
print "Difference between highest and second highest value", (probability_array[0][-1])*100 - (probability_array[0][-2])*100, "%"

# Number of instances per class in testing set  
np.histogram(y_test)

# Look where the value should be class 11 (potted plant)
np.where(y_test == 11)

# For these values, check the probability 
probability_array[2209993]

for i in range(0,11):
    print "Max. value for class " +str(i+1)+ ":", probability_array[:,i].max()*100, "%"


probability_array[:,0].shape
