==> 11_CreatePCtoVisualise.ipynb <==
# This creates my point cloud that I will later visualise in CloudCompare 

import pandas as pd
import os
import numpy as np
import sklearn 
import time
from sklearn.externals import joblib
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import f1_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix

# Define a data frame with all my data  
FILE_PATH = r"../DATA"
FILE_PATH_ScenarioB = r"../DATA/ML_datasets/ScenarioB"

y_test = np.loadtxt(FILE_PATH_ScenarioB+'/y_test_withXYZ.txt', delimiter=',')
X_test = np.loadtxt(FILE_PATH_ScenarioB+'/X_test_withXYZ.txt', delimiter=',')

y_train = np.loadtxt(FILE_PATH_ScenarioB+'/y_train_3000samples.txt', delimiter=',')
X_train = np.loadtxt(FILE_PATH_ScenarioB+'/X_train_3000samples.txt', delimiter=',')

X_test_classifier = X_test[:,3:]
print X_test[0]
print X_test_classifier[0]

start = time.time()
clf = RandomForestClassifier(max_depth=12, random_state=42, n_estimators=100, criterion='gini')
print 'Created Random Forest in:', float(time.time()-start), 'seconds'
start = time.time()
clf.fit(X_train, y_train)
print 'Fit model in:', float(time.time()-start), 'seconds'

# Make predictions for the unseen dataset 
predictions = clf.predict(X_test_classifier)

predictions.shape

y_test.shape

f1 = f1_score(y_test, predictions, average="weighted")

f1

# Now I want to export the true test labels y_test and the predictions together
# with the (xyz) coordinates

coordinates = X_test[:,0:3]

coordinates[0]

d = pd.DataFrame({'X':coordinates[:,0],'Y':coordinates[:,1],'Z':coordinates[:,2],
                 'true_labels':y_test, 'predictions':predictions})

d.shape

# Now export datadrame 
# Save file
d.to_csv(FILE_PATH_ScenarioB+'/VisualisePCdataset.txt', index=False)
